+++ integration_test/test/hls_endpoint_test.exs
@@ -72,7 +72,7 @@ defmodule Membrane.RTC.HLSEndpointTest do
       check_separate_hls_playlist(output_dir, 2, 2)
     end
 
-    test "creates correct hls stream with manual track addition", %{
+    test "creates correct hls stream with manual endpoint addition", %{
       rtc_engine: rtc_engine,
       tmp_dir: tmp_dir
     } do
@@ -104,14 +104,21 @@ defmodule Membrane.RTC.HLSEndpointTest do
                      },
                      @tracks_added_delay
 
-      assert_receive %Message.TrackAdded{track_id: track_id}
+      assert_receive %Message.TrackAdded{endpoint_id: ^file_endpoint_id, track_id: track_id}
 
       refute_receive {:trace, _pid, :call,
                       {Membrane.RTC.Engine, :subscribe,
                        [^rtc_engine, "hls-endpoint", ^track_id, _opts]}},
                      @tracks_added_delay
 
-      Engine.message_endpoint(rtc_engine, hls_endpoint_id, {:subscribe, [track_id]})
+      Engine.message_endpoint(rtc_engine, hls_endpoint_id, {:subscribe, [file_endpoint_id]})
+
+      # Checks if hls won't subscribe twice on the same track
+      # Should ignore tracks that is already subscribed for
+      Engine.message_endpoint(rtc_engine, hls_endpoint_id, {:subscribe, [file_endpoint_id]})
+
+      # Checks if hls won't crash if not existing endpoint is passed
+      Engine.message_endpoint(rtc_engine, hls_endpoint_id, {:subscribe, ["wrong_id"]})
 
       assert_receive {:trace, _pid, :call,
                       {Membrane.RTC.Engine, :subscribe,