+++ hls/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Update to Membrane Core 1.0 [#331](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/331)
 * Update Membrane Video Compositor dependency [#338](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/338)
 * Fix bug in Reques Handler [#341](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/341)
+* Manual subscription for endpoints [#343](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/343)
 
 ## 0.3.0
 * Add recording layout [#335](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/335)